Move mlflow logic into MlFlowReporter
Removes 'enable_mlflow' from local_build as this is part
of the Machine config itself. Furthermore, the config logic
of 'runtime.remote_logging.enable' remains. The workflow
generator will inject the MlFlowReporter given this flag.
