+++ gordo/builder/local_build.py
@@ -12,7 +12,7 @@ from gordo.machine import Machine
 
 
 def local_build(
-    config_str: str, enable_mlflow: bool = True, workspace_kwargs: dict = {}
+    config_str: str,
 ) -> Iterable[Tuple[Union[BaseEstimator, None], Machine]]:
     """
     Build model(s) from a bare Gordo config file locally.
@@ -26,15 +26,6 @@ def local_build(
     ----------
     config_str: str
         The raw yaml config file in string format.
-    enable_mlflow: bool
-        Flag to enable Mlflow logging of model building results. With `enable_mlflow`
-        set to `True`, passing `workspace_kwargs` an empty dict (default) will result in
-        local MLflow logging, while passing a dict of keyword arguments as defined in
-        :func:`~gordo.builder.mlflow_utils.get_mlflow_client` will result in
-        remote logging.
-    workspace_kwargs: dict
-        AzureML Workspace configuration to use for remote MLFlow tracking. See
-        :func:`gordo.builder.azure_utils.get_mlflow_client`.
 
     Examples
     --------
@@ -65,7 +56,7 @@ def local_build(
     ...                     estimator: sklearn.linear_model.base.LinearRegression
     ...         name: crazy-sweet-name
     ... '''
-    >>> models_n_metadata = local_build(config, enable_mlflow=False)
+    >>> models_n_metadata = local_build(config)
     >>> assert len(list(models_n_metadata)) == 1
 
     Returns
@@ -73,19 +64,8 @@ def local_build(
     Iterable[Tuple[Union[BaseEstimator, None], Machine]]
         A generator yielding tuples of models and their metadata.
     """
-    from gordo.builder.mlflow_utils import mlflow_context, log_machine
 
     config = get_dict_from_yaml(io.StringIO(config_str))
     normed = NormalizedConfig(config, project_name="local-build")
     for machine in normed.machines:
-        model, machine = ModelBuilder(machine=machine).build()
-
-        if enable_mlflow:
-            # This will enforce a single interactive login and automatically
-            # generate a new key for each model built
-            with mlflow_context(
-                name=machine.name, workspace_kwargs=workspace_kwargs
-            ) as (mlflow_client, run_id):
-                log_machine(mlflow_client, run_id, machine)
-
-        yield model, machine
+        yield ModelBuilder(machine=machine).build()
